server/requirements.txt
langgraph
langchain_groq 
langchain_core 
python_dotenv 
fastapi 
langchain_community